[VolumeControl] Optimize and enhance volume controls (#3561)
* [VolumeControl] Optimize and enhance volume controls

- Move all the volume control code into the VolumeControl.py screen code.
- Move the volume adjust functionality into the VolumeControl.py module.
- Remove the defunct VolumeAdjust.py module.
- Streamline volume control initialization in StartEnigma.py.
- Improve logging of all enhanced volume control functionality.
- Add option to show the volume level UI after adjustment changes are made.

This changes allows a tighter integration and co-operation between all of the volume control functions.

Volume adjust functionality now allows users to remember the last volume level used on each service in addition to the previous ability to define volume offsets for various services.

All services in both offset and last used modes are now listed in the Volume Adjust UI where they can be edited and/or deleted.

It is hoped that these changes will eliminate the need to use external plugins for volume management.  If any functionality has been missed or additional functionality is requested then please create an issue in GitHub.

---------

Co-authored-by: Ian <IanSav@users.noreply.github.com>